refactor: 자동차 이름 배열로 저장하도록 수정
