+++ src/main/java/racingcar/model/Car.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.model;
 
 public class Car {
