+++ src/main/java/racingcar/util/Converter.java
@@ -5,10 +5,10 @@ import java.util.List;
 
 public class Converter {
 
-    public List<String> parseCarList(String input) {
+    public String[] parseCarList(String input) {
         // 유효성 검사 추가 필요
         String[] inputArr = input.split(",");
-        return Arrays.asList(inputArr);
+        return inputArr;
     }
 
     public int convertTryCnt(String input) {
